refactor(monitor): split stat classes into another file
