+++ src/monitor/_common.py
@@ -1,0 +1,7 @@
+from typing import Final
+
+from ..log import getLogger
+
+TIMEOUT: Final[int] = 10 * 60  # 10 minutes
+
+logger = getLogger('RSStT.monitor')
